+++ Simulator/Vertices/Neuro/AllIFNeurons.cpp
@@ -142,7 +142,7 @@ void AllIFNeurons::createNeuron(int i, Layout &layout)
 
       default:
          LOG4CPLUS_DEBUG(vertexLogger_,
-                         "ERROR: unknown neuron type: " << static_cast<int>(layout.vertexTypeMap_[i]) << "@" << i);
+                         "ERROR: unknown neuron type: " << layout.vertexTypeMap_[i] << "@" << i);
          assert(false);
          break;
    }
